feat(datetime): allow input datetime string
